use flow in step-1,2. removing save_ld_json [WIP BROKEN COMMIT]

+++ spectral_trend_database/runner.py
@@ -121,21 +121,26 @@ def destination_strings(
 
 
 def table_name_and_paths(
-        table_name: str,
         *path_parts: str,
+        table_name: Optional[str] = None,
         year: Optional[int] = None,
-        ext: str = 'json'):
-    file_name = table_name.lower()
+        file_name: Optional[str] = None,
+        ext: Optional[str] = 'json'):
+    if table_name:
+        table_name = table_name.upper()
+    if not file_name:
+        file_name = table_name.lower()
     if year is not None:
         file_name = f'{file_name}-{year}'
-    file_name = f'{file_name}.{ext}'
+    if ext:
+        file_name = f'{file_name}.{ext}'
     local_dest = paths.local(
         *path_parts,
         file_name)
     gcs_dest = paths.gcs(
         *path_parts,
         file_name)
-    return table_name.upper(), local_dest, gcs_dest
+    return table_name, local_dest, gcs_dest
 
 
 def make_directories(*paths):
@@ -159,21 +164,30 @@ def save_to_gcp(
         bigquery_loc: str = c.LOCATION,
         gcs_uri: Optional[str] = None,
         remove_src: bool = False,
-        noisy: bool = NOISY):
+        noisy: bool = NOISY,
+        dry_run: bool = False) -> None:
     if gcs_dest:
         assert isinstance(src, str)
         assert isinstance(gcs_dest, str)
-        gcs_uri = gcp.upload_file(src, gcs_dest)
-        if noisy:
-            print('- uploaded to gcs:', gcs_uri)
-        if gcs_uri and remove_src:
-            Path(src).unlink()
+        if dry_run:
+            if noisy:
+                print('- dry_run [gcs]:', gcs_dest)
+        else:
+            gcs_uri = gcp.upload_file(src, gcs_dest)
+            if noisy:
+                print('- uploaded to gcs:', gcs_uri)
+            if gcs_uri and remove_src:
+                Path(src).unlink()
     if table_name:
-        if noisy:
-            print(f'- update table [{dataset_name}.{table_name}]')
-        assert isinstance(dataset_name, str)
-        assert isinstance(gcs_uri, str)
-        gcp.create_or_update_table_from_json(
-            gcp.load_or_create_dataset(dataset_name, bigquery_loc),
-            name=table_name,
-            uri=gcs_uri)
+        if dry_run:
+            if noisy:
+                print(f'- dry_run: update table [{dataset_name}.{table_name}]')
+        else:
+            if noisy:
+                print(f'- update table [{dataset_name}.{table_name}]')
+            assert isinstance(dataset_name, str)
+            assert isinstance(gcs_uri, str)
+            gcp.create_or_update_table_from_json(
+                gcp.load_or_create_dataset(dataset_name, bigquery_loc),
+                name=table_name,
+                uri=gcs_uri)
